+++ src/main/java/me/basiqueevangelist/pingspam/commands/GroupCommand.java
@@ -16,7 +16,7 @@ import me.basiqueevangelist.pingspam.PingSpam;
 import me.basiqueevangelist.pingspam.data.PingspamPlayerData;
 import me.basiqueevangelist.pingspam.network.ServerNetworkLogic;
 import me.basiqueevangelist.pingspam.utils.CommandUtil;
-import me.basiqueevangelist.pingspam.utils.NameLogic;
+import me.basiqueevangelist.pingspam.logic.NameLogic;
 import me.basiqueevangelist.pingspam.utils.NameUtil;
 import me.lucko.fabric.api.permissions.v0.Permissions;
 import net.minecraft.command.argument.GameProfileArgumentType;
