+++ src/main/java/me/basiqueevangelist/pingspam/commands/mail/SendCommand.java
@@ -9,8 +9,8 @@ import me.basiqueevangelist.onedatastore.api.DataStore;
 import me.basiqueevangelist.pingspam.PingSpam;
 import me.basiqueevangelist.pingspam.data.MailMessage;
 import me.basiqueevangelist.pingspam.data.PechkinPlayerData;
-import me.basiqueevangelist.pingspam.utils.IgnoreLogic;
-import me.basiqueevangelist.pingspam.utils.MailLogic;
+import me.basiqueevangelist.pingspam.logic.IgnoreLogic;
+import me.basiqueevangelist.pingspam.logic.MailLogic;
 import me.basiqueevangelist.pingspam.utils.CommandUtil;
 import me.lucko.fabric.api.permissions.v0.Permissions;
 import net.minecraft.command.argument.GameProfileArgumentType;
