+++ src/main/java/me/basiqueevangelist/pingspam/mixin/ServerPlayNetworkHandlerMixin.java
@@ -2,7 +2,7 @@ package me.basiqueevangelist.pingspam.mixin;
 
 import me.basiqueevangelist.onedatastore.api.DataStore;
 import me.basiqueevangelist.pingspam.PingSpam;
-import me.basiqueevangelist.pingspam.utils.GroupChatLogic;
+import me.basiqueevangelist.pingspam.logic.GroupChatLogic;
 import net.minecraft.network.ClientConnection;
 import net.minecraft.network.message.SignedMessage;
 import net.minecraft.server.MinecraftServer;
@@ -10,7 +10,6 @@ import net.minecraft.server.network.ConnectedClientData;
 import net.minecraft.server.network.ServerCommonNetworkHandler;
 import net.minecraft.server.network.ServerPlayNetworkHandler;
 import net.minecraft.server.network.ServerPlayerEntity;
-import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
